LayoutTests/imported/w3c/web-platform-tests/scroll-animations/css/view-timeline-animation-expected.txt


FAIL Default view-timeline assert_equals: expected "-1" but got "100"
FAIL Horizontal view-timeline assert_equals: expected "-1" but got "100"
FAIL Multiple view-timelines on the same element assert_equals: expected "-1" but got "100"
